refactor: making select an independent module
